+++ trunk/YUVToolkit/ColorMap.h
@@ -4,6 +4,6 @@
 #include "YT_Interface.h"
 
 
-void CreateColorMap(FramePtr frame, DistMapPtr distMap, int width, int height, double upperRange, double lowerRange, bool biggerValueIsBetter);
+void CreateColorMap(FramePtr frame, DistMapPtr distMap, int width, int height, float upperRange, float lowerRange, bool biggerValueIsBetter);
 
 #endif
